requirements.txt
Django==3.2.7
requests==2.26.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0